set int value for limit

+++ extend.php
@@ -25,7 +25,7 @@ return [
         ->js(__DIR__.'/js/dist/admin.js')
         ->css(__DIR__.'/resources/less/admin.less'),
     (new Extend\ApiController(AbstractSerializeController::class))
-        ->setLimit(resolve(SettingsRepositoryInterface::class)->get('block-cat.article_per_page')),
+        ->setLimit((int)resolve(SettingsRepositoryInterface::class)->get('block-cat.article_per_page')),
     (new Extend\Settings())
         -> serializeToForum('block-cat.articlePerPage', 'block-cat.article_per_page')
         -> serializeToForum('block-cat.paginationPosition', 'block-cat.pagination_position'),
